# Improves Pyroscope's scrape sync