This improves the sync method of Pyroscope scrapes based on some investigation.

The change is basically caching values of labels, url and hash of targets to avoid regenerating them when we sync and scrape profiles.

To ensure we're improving something I've added a benchmark.

Result:

```
benchstat before.txt after.txt
name     old time/op    new time/op    delta
Sync-16     796µs ±13%     167µs ± 3%  -79.06%  (p=0.008 n=5+5)

name     old alloc/op   new alloc/op   delta
Sync-16     609kB ± 0%     162kB ± 0%  -73.35%  (p=0.008 n=5+5)

name     old allocs/op  new allocs/op  delta
Sync-16     8.26k ± 0%     1.55k ± 0%  -81.29%  (p=0.008 n=5+5)
```

I'm sure there's more we can do.