@thampiotr 2023-10-20
Merging this as it provides a lot of value by addressing the core problem. We can follow-up with more clean up work afterwards.